+++ cli/src/main.rs
@@ -113,7 +113,7 @@ async fn main() {
                             .subcommand(
                                 Command::new("files")
                                 .about("Configures a Files output which will store events on disk at configured path")
-                                .arg(arg!(<path> "Destination path that can use variables. Example: \"/archive/{ip}/{principal}/{node}/messages\", where {ip} is the string representation of the IP addr of the machine and {principal} its Kerberos principal. See documentation for other variables."))
+                                .arg(arg!(<path> "Destination path that can use variables. Example: \"/archive/{ip}/{client}/{node}/messages\", where {ip} is the string representation of the IP addr of the machine and {client} its identifier (Kerberos Principal or TLS subject). See documentation for other variables."))
                             )
                             .subcommand(
                                 Command::new("unixdatagram")
@@ -137,35 +137,6 @@ async fn main() {
                             .about("Disables an output of this subscription")
                             .arg(arg!(-y --yes "Do not prompt for confirmation"))
                             .arg(arg!(<index> "Index of the output to disable").value_parser(value_parser!(usize)))
-                        )
-                    )
-                    .subcommand(
-                        Command::new("filter")
-                        .about("Manage the principals filter of a subscription.")
-                        .subcommand(
-                            Command::new("set")
-                            .about("Set the principals filter.")
-                            .arg(arg!(<operation> "Possible operations are: 'none' (don't filter), 'only' (only apply this subscription to the given principals), 'except' (apply this subcription to everyone except the given principals).").value_parser(["none", "only", "except"]).required(true))
-                            .arg(arg!(<principals> ... "Principals to filter. The comparison of principal names is case-sensitive.").action(ArgAction::Append).required(false))
-                        )
-                        .subcommand(
-                            Command::new("princs")
-                            .about("Manage the principals of the filter.")
-                            .subcommand(
-                                Command::new("add")
-                                .about("Add a principal (case-sensitive)")
-                                .arg(arg!(<principal> "Principal to add. The comparison of principal names is case-sensitive."))
-                            )
-                            .subcommand(
-                                Command::new("delete")
-                                .about("Delete a principal (case-sensitive)")
-                                .arg(arg!(<principal> "Principal to delete. The comparison of principal names is case-sensitive."))
-                            )
-                            .subcommand(
-                                Command::new("set")
-                                .about("Set the principals (case-sensitive)")
-                                .arg(arg!(<principals> ... "Principals to filter. The comparison of principal names is case-sensitive.").action(ArgAction::Append).required(true))
-                            )
                         )
                     )
                     .arg(arg!(-q --query <QUERY> "File containing the query (XML format)"))
@@ -263,12 +234,13 @@ async fn main() {
                     .arg(arg!(<subscription> "Name or UUID of the subscription"))
                 )
                 .subcommand(
-                    Command::new("machines")
-                    .about("Show subscribing machines. Defaults to all machines ever seen.")
+                    Command::new("clients")
+                    .alias("machines")
+                    .about("Show subscribing clients. Defaults to all clients ever seen.")
                     .arg(arg!(<subscription> "Name or UUID of the subscription"))
-                    .arg(arg!(-a --active "Only show active machines"))
-                    .arg(arg!(-l --alive "Only show machines that are alive but not active"))
-                    .arg(arg!(-d --dead "Only show dead machines principal"))
+                    .arg(arg!(-a --active "Only show active clients"))
+                    .arg(arg!(-l --alive "Only show clients that are alive but not active"))
+                    .arg(arg!(-d --dead "Only show dead clients"))
                     .arg(
                         arg!(-i --interval <INTERVAL> "Duration after which a machine is considered alive if no events are received or dead if no heartbeats are received. \
                             Defaults to heartbeat-interval")